requirements.txt
requests                          
langchain                         
sentence-transformers            
chromadb                          
pinecone-client                    


pdfplumber                      


arize                             


pytest                            